pyproject.toml
[project]
name = "shop-tools"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adrf>=0.1.11",
    "boto3>=1.40.63",
    "celery[redis]>=5.5.3",
    "channels>=4.3.1",
    "colorlog>=6.10.1",
    "django>=5.2.7",
    "django-async-orm>=0.1.12",
    "django-celery-beat>=2.8.1",
    "django-ckeditor-5>=0.2.18",
    "django-cors-headers>=4.9.0",
    "django-daterangefilter>=1.0.1",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-json-widget>=2.1.0",
    "django-redis>=6.0.0",
    "django-silk>=5.4.3",
    "django-storages>=1.14.6",
    "django-treebeard>=4.7.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "ipython>=9.6.0",
    "locust>=2.42.1",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest-asyncio>=1.2.0",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
]

[tool.ruff]
line-length = 88
